+++ src/api/accountsApi.js
@@ -2,6 +2,7 @@ import api from "./axios.js";
 import { toast } from "react-toastify";
 
 const constroller = "/accounts";
+const balanceEndpoint = "/balance";
 const accountByCBU = "/cbu/";
 const toastOptions = {
   position: "top-center",
@@ -23,13 +24,25 @@ export default class AccountsApi {
             //refrescar token
             toast.error(
               "Su sesión ha expirado, por favor vuelva a iniciar sesión",
-              toastOptions,
+              toastOptions
             );
           } else {
             toast.error(error.response.data.message, toastOptions);
           }
           reject(error);
         });
+    });
+  }
+
+  static async balance() {
+    return api.get(constroller + balanceEndpoint).then((response) => {
+      console.log(response);
+    });
+  }
+
+  static async accountInfo(id) {
+    return api.get(constroller + `/${id}`).catch((error) => {
+      console.log(error);
     });
   }
 }
